Synthesis: refactor the modules
This commit removes the use of static methods and replaces it by inheritance.

This splits up the former SynthesizeModule into an AbstractSynthesizeModule,
which contains the common methods, classes and interfaces, and the real
SynthesizeModule, which now only contains things which are specific to this
module.
