+++ src/module/uniol/apt/analysis/synthesize/SynthesizeModule.java
@@ -19,7 +19,6 @@
 
 package uniol.apt.analysis.synthesize;
 
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
@@ -28,14 +27,9 @@ import java.util.Set;
 import java.util.TreeMap;
 import java.util.TreeSet;
 
-import uniol.apt.adt.pn.PetriNet;
-import uniol.apt.adt.pn.Place;
 import uniol.apt.adt.ts.State;
 import uniol.apt.adt.ts.TransitionSystem;
-import uniol.apt.analysis.synthesize.separation.SeparationUtility;
-import uniol.apt.module.AbstractModule;
 import uniol.apt.module.AptModule;
-import uniol.apt.module.Category;
 import uniol.apt.module.Module;
 import uniol.apt.module.ModuleInput;
 import uniol.apt.module.ModuleInputSpec;
@@ -48,7 +42,7 @@ import uniol.apt.module.exception.ModuleException;
  * @author Uli Schlachter
  */
 @AptModule
-public class SynthesizeModule extends AbstractModule implements Module {
+public class SynthesizeModule extends AbstractSynthesizeModule implements Module {
 
 	@Override
 	public String getShortDescription() {
@@ -58,7 +52,7 @@ public class SynthesizeModule extends AbstractModule implements Module {
 	@Override
 	public String getLongDescription() {
 		return getShortDescription() + ".\n\n"
-			+ getOptionsDescription() + "\n\nExample calls:\n\n"
+			+ getOptionsDescription("", "") + "\n\nExample calls:\n\n"
 			+ " apt " + getName() + " none lts.apt\n"
 			+ " apt " + getName() + " 3-bounded lts.apt\n"
 			+ " apt " + getName() + " pure,safe lts.apt\n"
@@ -70,137 +64,16 @@ public class SynthesizeModule extends AbstractModule implements Module {
 		return "synthesize";
 	}
 
-	static public String getOptionsDescription() {
-		return getOptionsDescription("", "");
-	}
-
-	static public String getOptionsDescription(String extraOptions, String extraOptionsDescriptions) {
-		return "Supported options are: none, [k]-bounded, safe, pure, plain, tnet,"
-			+ " generalized-marked-graph (gmg), marked-graph (mg), generalized-output-nonbranching (gon)"
-			+ " output-nonbranching (on), conflict-free (cf), homogeneous,"
-			+ " upto-language-equivalence (language, le), " + extraOptions + "minimize (minimal), verbose "
-			+ "and quick-fail.\n\nThe meaning of these options is as follows:\n"
-			+ " - none: No further requirements are made.\n"
-			+ " - [k]-bounded: In every reachable marking, every place contains at most [k] tokens.\n"
-			+ " - safe: Equivalent to 1-bounded.\n"
-			+ " - pure: Every transition either consumes or produces tokens on a place, but noth both"
-			+ " (=no side-conditions).\n"
-			+ " - plain: Every flow has a weight of at most one.\n"
-			+ " - tnet: Every place's preset and postset contains at most one entry.\n"
-			+ " - generalized-marked-graph: Every place's preset and postset contains exactly one entry.\n"
-			+ " - marked-graph: generalized-marked-graph + plain.\n"
-			+ " - generalized-output-nonbranching: Every place's postset contains at most one entry.\n"
-			+ " - output-nonbranching: generalized-output-nonbranching + plain.\n"
-			+ " - conflict-free: The Petri net is plain and every place either has at most one entry in"
-			+ " its postset or its preset is contained in its postset.\n"
-			+ " - homogeneous: All outgoing flows from a place have the same weight.\n"
-			+ " - minimize: The Petri net has as few places as possible.\n"
-			+ extraOptionsDescriptions
-			+ "The following options only affect the output, but not the produced Petri net:\n"
-			+ " - verbose: Print details about each calculated region/place.\n"
-			+ " - quick-fail: Stop the algorithm when the result 'success: No' is clear.";
-	}
-
-	static public void requireCommon(ModuleInputSpec inputSpec) {
-		inputSpec.addParameter("options", String.class, "Comma separated list of options");
-	}
-
 	@Override
-	public void require(ModuleInputSpec inputSpec) {
-		requireCommon(inputSpec);
+	protected void requireExtra(ModuleInputSpec inputSpec) {
 		inputSpec.addParameter("lts", TransitionSystem.class,
 				"The LTS that should be synthesized to a Petri Net");
 	}
 
-	static public void provideCommon(ModuleOutputSpec outputSpec) {
-		outputSpec.addReturnValue("success", Boolean.class, ModuleOutputSpec.PROPERTY_SUCCESS);
-		outputSpec.addReturnValue("solvedEventStateSeparationProblems", String.class);
-		outputSpec.addReturnValue("pn", PetriNet.class,
-			ModuleOutputSpec.PROPERTY_FILE, ModuleOutputSpec.PROPERTY_RAW);
-	}
-
 	@Override
-	public void provide(ModuleOutputSpec outputSpec) {
-		provideCommon(outputSpec);
+	protected void provideExtra(ModuleOutputSpec outputSpec) {
 		outputSpec.addReturnValue("failedStateSeparationProblems", String.class);
 		outputSpec.addReturnValue("failedEventStateSeparationProblems", String.class);
-	}
-
-	static protected interface TransitionSystemForOptions {
-		public Collection<String> supportedExtraOptions();
-		public TransitionSystem getTS(Collection<String> enabledOptions);
-	}
-
-	static protected class ReturnTS implements TransitionSystemForOptions {
-		final private TransitionSystem ts;
-
-		public ReturnTS(TransitionSystem ts) {
-			this.ts = ts;
-		}
-
-		@Override
-		public Collection<String> supportedExtraOptions() {
-			return Collections.emptyList();
-		}
-
-		@Override
-		public TransitionSystem getTS(Collection<String> enabledOptions) {
-			return ts;
-		}
-	}
-
-	static public SynthesizePN runSynthesis(TransitionSystemForOptions tsForOpts, ModuleInput input,
-				ModuleOutput output)
-			throws ModuleException {
-		String quickFailStr = "quick-fail", verboseStr = "verbose";
-		Collection<String> languageEquivalenceStr = Arrays.asList("upto-language-equivalence", "language",
-				"le");
-		Collection<String> minimizeStr = Arrays.asList("minimize", "minimise", "minimal");
-		Set<String> supportedExtraOptions = new HashSet<>(Arrays.asList(quickFailStr, verboseStr));
-		supportedExtraOptions.addAll(languageEquivalenceStr);
-		supportedExtraOptions.addAll(minimizeStr);
-		supportedExtraOptions.addAll(tsForOpts.supportedExtraOptions());
-
-		Options options = Options.parseProperties(input.getParameter("options", String.class),
-				supportedExtraOptions);
-		boolean quickFail = options.extraOptions.contains(quickFailStr);
-		boolean verbose = options.extraOptions.contains(verboseStr);
-		boolean languageEquivalence = !Collections.disjoint(options.extraOptions, languageEquivalenceStr);
-		boolean minimize = !Collections.disjoint(options.extraOptions, minimizeStr);
-
-		SynthesizePN synthesize;
-		SynthesizePN.Builder builder = new SynthesizePN.Builder(tsForOpts.getTS(options.extraOptions))
-			.setProperties(options.properties)
-			.setQuickFail(quickFail);
-		if (languageEquivalence)
-			synthesize = builder.buildForLanguageEquivalence();
-		else
-			synthesize = builder.buildForIsomorphicBehavior();
-
-		PetriNet pn;
-		Collection<Region> regions;
-		if (synthesize.wasSuccessfullySeparated() && minimize) {
-			MinimizePN min = new MinimizePN(synthesize);
-			regions = min.getSeparatingRegions();
-			pn = min.synthesizePetriNet();
-		} else {
-			regions = synthesize.getSeparatingRegions();
-			pn = synthesize.synthesizePetriNet();
-		}
-		if (pn != null)
-			for (Place p : pn.getPlaces())
-				p.removeExtension(Region.class.getName());
-
-		output.setReturnValue("success", Boolean.class, synthesize.wasSuccessfullySeparated());
-		output.setReturnValue("pn", PetriNet.class, pn);
-		if (verbose) {
-			output.setReturnValue("solvedEventStateSeparationProblems", String.class,
-					getSolvedEventStateSeparationProblems(
-						synthesize.getUtility().getTransitionSystem(),
-						regions));
-		}
-
-		return synthesize;
 	}
 
 	@Override
@@ -228,163 +101,24 @@ public class SynthesizeModule extends AbstractModule implements Module {
 		}
 	}
 
-	@Override
-	public Category[] getCategories() {
-		return new Category[]{Category.LTS};
-	}
+	static protected class ReturnTS implements TransitionSystemForOptions {
+		final private TransitionSystem ts;
 
-	/**
-	 * Get the IDs of a set of states.
-	 * @param states The states to look at
-	 * @return The set of state IDs.
-	 */
-	static public Set<String> getStateIDs(Set<State> states) {
-		Set<String> result = new TreeSet<>();
-		for (State state : states)
-			result.add(state.getId());
-		return result;
-	}
-
-	/**
-	 * Get a string representation that describes for each region which Event/State Separation Problems it solves.
-	 * @param ts The transition system that should be solved.
-	 * @param regions The regions that solve (part of) the transition system.
-	 * @return A string representation of the solved problems.
-	 */
-	static public String getSolvedEventStateSeparationProblems(TransitionSystem ts, Collection<Region> regions) {
-		StringBuilder result = new StringBuilder("");
-
-		for (Region region : regions) {
-			result.append("\nRegion ").append(region).append(":");
-			for (String event : ts.getAlphabet()) {
-				Set<State> states = new HashSet<>();
-				for (State state : ts.getNodes()) {
-					if (!SeparationUtility.isEventEnabled(state, event)
-							&& SeparationUtility.isSeparatingRegion(region, state, event))
-						states.add(state);
-				}
-				if (!states.isEmpty()) {
-					result.append("\n\tseparates event ");
-					result.append(event);
-					result.append(" at states ");
-					result.append(getStateIDs(states));
-				}
-			}
+		public ReturnTS(TransitionSystem ts) {
+			this.ts = ts;
 		}
 
-		String ret = result.toString();
-		if (ret.isEmpty())
-			return "none";
-		return ret;
-	}
-
-	/**
-	 * Instances of this class hold options that can be specified for synthesis.
-	 */
-	static public class Options {
-		public final PNProperties properties;
-		public final Set<String> extraOptions;
-
-		public Options(PNProperties properties, Collection<String> extraOptions) {
-			this.properties = properties;
-			this.extraOptions = Collections.unmodifiableSet(new HashSet<>(extraOptions));
+		@Override
+		public Collection<String> supportedExtraOptions() {
+			return Collections.emptyList();
 		}
 
-		/**
-		 * Parse the given string into an Options instance
-		 * @param properties the string to parse
-		 * @return A representation of the requested options
-		 * @throws ModuleException if the properties string is malformed
-		 */
-		static public Options parseProperties(String properties) throws ModuleException {
-			return parseProperties(properties, Collections.<String>emptySet());
-		}
-
-		/**
-		 * Parse the given string into an Options instance
-		 * @param properties the string to parse
-		 * @param supportedExtraOptions extra options which should also be supported
-		 * @return A representation of the requested options
-		 * @throws ModuleException if the properties string is malformed
-		 */
-		static public Options parseProperties(String properties, Collection<String> supportedExtraOptions)
-				throws ModuleException {
-			PNProperties result = new PNProperties();
-
-			// Explicitly allow empty string
-			properties = properties.trim();
-			if (properties.isEmpty())
-				return new Options(result, Collections.<String>emptySet());
-
-			Set<String> extraOptions = new HashSet<>();
-			for (String prop : properties.split(",")) {
-				prop = prop.trim().toLowerCase();
-
-				if (supportedExtraOptions.contains(prop)) {
-					extraOptions.add(prop);
-					continue;
-				}
-
-				switch (prop) {
-					case "none":
-						break;
-					case "safe":
-						result = result.requireSafe();
-						break;
-					case "pure":
-						result = result.setPure(true);
-						break;
-					case "plain":
-						result = result.setPlain(true);
-						break;
-					case "tnet":
-						result = result.setTNet(true);
-						break;
-					case "generalized-marked-graph":
-					case "gmg":
-						result = result.setMarkedGraph(true);
-						break;
-					case "marked-graph":
-					case "mg":
-						result = result.setMarkedGraph(true).setPlain(true);
-						break;
-					case "generalized-output-nonbranching":
-					case "gon":
-						result = result.setOutputNonbranching(true);
-						break;
-					case "output-nonbranching":
-					case "on":
-						result = result.setOutputNonbranching(true).setPlain(true);
-						break;
-					case "conflict-free":
-					case "cf":
-						result = result.setConflictFree(true);
-						break;
-					case "homogeneous":
-						result = result.setHomogeneous(true);
-						break;
-					default:
-						if (prop.endsWith("-bounded")) {
-							String value = prop.substring(0, prop.length() - "-bounded".length());
-							int k;
-							try {
-								k = Integer.parseInt(value);
-							} catch (NumberFormatException e) {
-								throw new ModuleException("Cannot parse '" + prop + "': "
-										+ "Invalid number for property 'k-bounded'");
-							}
-							if (k < 1)
-								throw new ModuleException("Cannot parse '" + prop + "': "
-										+ "Bound must be positive");
-							result = result.requireKBounded(k);
-						} else
-							throw new ModuleException("Cannot parse '" + prop
-									+ "': Unknown property");
-				}
-			}
-			return new Options(result, extraOptions);
+		@Override
+		public TransitionSystem getTS(Collection<String> enabledOptions) {
+			return ts;
 		}
 	}
+
 }
 
 // vim: ft=java:noet:sw=8:sts=8:ts=8:tw=120
